+++ videoanalyst/model/loss/loss_impl/focal_loss.py
@@ -7,100 +7,8 @@ from torch import nn, tensor
 
 from ...module_base import ModuleBase
 from ..loss_base import TRACK_LOSSES
+from ...common_opr.common_loss import sigmoid_focal_loss_jit
 
-
-def sigmoid_focal_loss(
-    inputs: torch.Tensor,
-    targets: torch.Tensor,
-    alpha: float = -1,
-    gamma: float = 2,
-    reduction: str = "none",
-) -> torch.Tensor:
-    """
-    Loss used in RetinaNet for dense detection: https://arxiv.org/abs/1708.02002.
-    Args:
-        inputs: A float tensor of arbitrary shape.
-                The predictions for each example.
-        targets: A float tensor with the same shape as inputs. Stores the binary
-                 classification label for each element in inputs
-                (0 for the negative class and 1 for the positive class).
-        alpha: (optional) Weighting factor in range (0,1) to balance
-                positive vs negative examples. Default = -1 (no weighting).
-        gamma: Exponent of the modulating factor (1 - p_t) to
-               balance easy vs hard examples.
-        reduction: 'none' | 'mean' | 'sum'
-                 'none': No reduction will be applied to the output.
-                 'mean': The output will be averaged.
-                 'sum': The output will be summed.
-    Returns:
-        Loss tensor with the reduction option applied.
-    """
-    p = torch.sigmoid(inputs)
-    ce_loss = F.binary_cross_entropy_with_logits(
-        inputs, targets, reduction="none"
-    )
-    p_t = p * targets + (1 - p) * (1 - targets)
-    loss = ce_loss * ((1 - p_t) ** gamma)
-
-    if alpha >= 0:
-        alpha_t = alpha * targets + (1 - alpha) * (1 - targets)
-        loss = alpha_t * loss
-    if reduction == "mean":
-        loss = loss.mean()
-    elif reduction == "sum":
-        loss = loss.sum()
-
-    return loss
-
-
-sigmoid_focal_loss_jit = torch.jit.script(
-    sigmoid_focal_loss
-)  # type: torch.jit.ScriptModule
-
-
-def sigmoid_focal_loss_star(
-    inputs: torch.Tensor,
-    targets: torch.Tensor,
-    alpha: float = -1,
-    gamma: float = 1,
-    reduction: str = "none",
-) -> torch.Tensor:
-    """
-    FL* described in RetinaNet paper Appendix: https://arxiv.org/abs/1708.02002.
-    Args:
-        inputs: A float tensor of arbitrary shape.
-                The predictions for each example.
-        targets: A float tensor with the same shape as inputs. Stores the binary
-                 classification label for each element in inputs
-                (0 for the negative class and 1 for the positive class).
-        alpha: (optional) Weighting factor in range (0,1) to balance
-                positive vs negative examples. Default = -1 (no weighting).
-        gamma: Gamma parameter described in FL*. Default = 1 (no weighting).
-        reduction: 'none' | 'mean' | 'sum'
-                 'none': No reduction will be applied to the output.
-                 'mean': The output will be averaged.
-                 'sum': The output will be summed.
-    Returns:
-        Loss tensor with the reduction option applied.
-    """
-    shifted_inputs = gamma * (inputs * (2 * targets - 1))
-    loss = -F.logsigmoid(shifted_inputs) / gamma
-
-    if alpha >= 0:
-        alpha_t = alpha * targets + (1 - alpha) * (1 - targets)
-        loss *= alpha_t
-
-    if reduction == "mean":
-        loss = loss.mean()  # pyre-ignore
-    elif reduction == "sum":
-        loss = loss.sum()  # pyre-ignore
-
-    return loss
-
-
-sigmoid_focal_loss_star_jit = torch.jit.script(
-    sigmoid_focal_loss_star
-)  # type: torch.jit.ScriptModule
 
 @TRACK_LOSSES.register
 class FocalLoss(ModuleBase):
@@ -152,9 +60,10 @@ class FocalLoss(ModuleBase):
         index_mat = vlabel.type(torch.LongTensor)
         onehot_ = zero_mat.scatter(2, index_mat, one_mat)
         onehot = onehot_[:, :, 1:].type(torch.Tensor).to(pred.device)
-        loss = sigmoid_focal_loss_jit(pred, onehot, self._hyper_params["alpha"], self._hyper_params["gamma"], "none")
+        loss = sigmoid_focal_loss_jit(pred, onehot, self._hyper_params["alpha"],
+                                      self._hyper_params["gamma"], "none")
         positive_mask = (label > 0).type(torch.Tensor).to(pred.device)
-        loss = (loss.sum(dim=2)*mask.squeeze(2)).sum()/(torch.max(positive_mask.sum(), self.t_one)) * self._hyper_params["weight"]
+        loss = (loss.sum(dim=2) * mask.squeeze(2)).sum() / (torch.max(
+            positive_mask.sum(), self.t_one)) * self._hyper_params["weight"]
         extra = dict()
         return loss, extra
-
